move validating bridge to the commons
