+++ apis/sgv2-docsapi/src/test/java/io/stargate/sgv2/docsapi/service/query/ReadBridgeServiceTest.java
@@ -28,11 +28,11 @@ import io.smallrye.mutiny.helpers.test.AssertSubscriber;
 import io.smallrye.mutiny.helpers.test.UniAssertSubscriber;
 import io.stargate.bridge.grpc.Values;
 import io.stargate.bridge.proto.QueryOuterClass;
+import io.stargate.sgv2.common.bridge.AbstractValidatingStargateBridgeTest;
+import io.stargate.sgv2.common.bridge.ValidatingStargateBridge;
 import io.stargate.sgv2.docsapi.DocsApiTestSchemaProvider;
 import io.stargate.sgv2.docsapi.api.properties.document.DocumentProperties;
 import io.stargate.sgv2.docsapi.api.v2.model.dto.ExecutionProfile;
-import io.stargate.sgv2.docsapi.bridge.AbstractValidatingStargateBridgeTest;
-import io.stargate.sgv2.docsapi.bridge.ValidatingStargateBridge;
 import io.stargate.sgv2.docsapi.service.ExecutionContext;
 import io.stargate.sgv2.docsapi.service.common.model.Paginator;
 import io.stargate.sgv2.docsapi.service.query.condition.BaseCondition;
