+++ apis/sgv2-docsapi/src/test/java/io/stargate/sgv2/docsapi/service/query/search/resolver/impl/InMemoryDocumentsResolverTest.java
@@ -27,11 +27,11 @@ import io.quarkus.test.junit.TestProfile;
 import io.smallrye.mutiny.helpers.test.AssertSubscriber;
 import io.stargate.bridge.grpc.Values;
 import io.stargate.bridge.proto.QueryOuterClass;
+import io.stargate.sgv2.common.bridge.AbstractValidatingStargateBridgeTest;
+import io.stargate.sgv2.common.bridge.ValidatingStargateBridge;
 import io.stargate.sgv2.docsapi.DocsApiTestSchemaProvider;
 import io.stargate.sgv2.docsapi.OpenMocksTest;
 import io.stargate.sgv2.docsapi.api.properties.document.DocumentProperties;
-import io.stargate.sgv2.docsapi.bridge.AbstractValidatingStargateBridgeTest;
-import io.stargate.sgv2.docsapi.bridge.ValidatingStargateBridge;
 import io.stargate.sgv2.docsapi.service.ExecutionContext;
 import io.stargate.sgv2.docsapi.service.common.model.Paginator;
 import io.stargate.sgv2.docsapi.service.query.FilterExpression;
