+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/common/bridge/AbstractValidatingStargateBridgeTest.java
@@ -10,10 +10,11 @@
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
+ *  See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
-package io.stargate.sgv2.docsapi.bridge;
+package io.stargate.sgv2.common.bridge;
 
 import static org.mockito.Mockito.when;
